+++ setup.cfg
@@ -15,4 +15,4 @@ keywords = LLM, LLaMA, GPT, Falcon
 python_requires = >=3.8
 package_dir =
     llm_inference = src/llm_inference
-    chatbot = src/chatbot
+    llm_chain = src/llm_chain
